Restructure project directory and add node-sass
